feat(billing): refactor billing module to improve organization and maintainability by consolidating utility functions and introducing a centralized client configuration refactor(billing): move Polar client initialization and related functions to a dedicated client module for better separation of concerns refactor(billing): create eligibility and subscription management utilities to streamline subscription operations and improve code clarity chore(billing): remove deprecated utils file and update imports across billing routes to use the new structure for better maintainability
