+++ app/api/billing/can-subscribe/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { auth, currentUser } from '@clerk/nextjs/server';
 import { ensureUserSynced } from '@/lib/user-sync';
-import { getUserSubscription, getSubscriptionEligibility } from '@/lib/billing/utils';
+import { getUserSubscription, getSubscriptionEligibility } from '@/lib/billing';
 import { ApiErrorHandler } from '@/lib/api/errors';
 
 export async function GET() {
