+++ app/api/billing/reactivate-subscription/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import { auth, currentUser } from '@clerk/nextjs/server';
-import { reactivateUserSubscription, getSubscriptionEligibility } from '@/lib/billing/utils';
+import { reactivateUserSubscription, getSubscriptionEligibility } from '@/lib/billing';
 import { getUserSubscriptionWithSync } from '@/lib/billing/polar-sync';
 import { ensureUserSynced } from '@/lib/user-sync';
 import { ApiErrorHandler } from '@/lib/api/errors';
