+++ app/api/billing/subscription-status/route.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from 'next/server';
 import { auth, currentUser } from '@clerk/nextjs/server';
 import { ensureUserSynced } from '@/lib/user-sync';
-import { getUserSubscription } from '@/lib/billing/utils';
+import { getUserSubscription } from '@/lib/billing';
 import { ApiErrorHandler } from '@/lib/api/errors';
 
 export async function GET() {
